LayoutTests/media/W3C/audio/readyState/readyState_initial.html

<!doctype html>
<html>
 <head>
  <title>audio.readyState - default state</title>
  <script src="../../w3cwrapper.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-networkstate">spec reference</a></p>
  <audio id="a">
  </audio>
  <div id="log"></div>
  <script>
test(function() {
 var a = document.getElementById("a");
 assert_equals(
  a.readyState,
  a.HAVE_NOTHING,
  "audioElement.readyState should be HAVE_NOTHING to begin with");
});
  </script>
 </body>
</html>
