LayoutTests/media/W3C/audio/readyState/readyState_initial-expected.txt

audio.readyState - default state

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "0" is "0"

TEST COMPLETE
spec reference
